+++ pm4py/algo/connectors/variants/__init__.py
@@ -15,4 +15,3 @@
     along with PM4Py.  If not, see <https://www.gnu.org/licenses/>.
 '''
 
-from pm4py.algo.connectors.variants import chrome_history, firefox_history, github_repo, outlook_calendar, outlook_mail_extractor, windows_events